2.0.8 - Fixes, searching (#11)
* 2.0.8 - Fixes, searching

* fixes

* wip

* fix

* legalize fathers

* wip

* fix!!

* wip

* exp

* wip

* fixes

* fixes

* wip

* wip

* wip

* wip

* wip

* wip

* warning fixes

* wip

* wip

* wip

* fixes and ollama chat

* fix

* wip

* wip

* searchapi

* wip

* renames

* RunCodeTask

* Fixed codingtask

* fix

* doc update

* interpreters

* fix

* fix

* fix

* fix

* fix
x

* fix?

+++ jo-penai/src/main/kotlin/com/simiacryptus/cognotik/audio/TranscriptionProcessor.kt
@@ -51,7 +51,7 @@ open class TranscriptionProcessor(
                     |Text: ${transcriptionResult.text}
                     |Prompt: ${transcriptionResult.prompt}
                     |Processing time: ${transcriptionResult.processingTime}ms
-                    |Audio duration: ${recordAudio.samples.size / recordAudio.audioFormat.sampleRate.toFloat()}s
+                    |Audio duration: ${recordAudio.samples.size / recordAudio.audioFormat.sampleRate}s
                     |""".trimMargin()
                 )
             }
